Begin gutting state_manip.ts
The module has grown to contain all sorts of unrelated things. Split
into a module containing the ContextInfo class, a module containing
the code to create context phrases, and a module with actual state
manipulation. Later, the state manipulation will stop using the
context info and will move to utils/thingtalk
