+++ lib/transaction-dialogues/dialogue_acts/action-confirm.ts
@@ -25,8 +25,8 @@ import { Ast, Type } from 'thingtalk';
 
 import * as C from '../../templates/ast_manip';
 
+import { ContextInfo } from '../context-info';
 import {
-    ContextInfo,
     makeAgentReply,
     makeSimpleState,
     setOrAddInvocationParam,
