+++ lib/transaction-dialogues/dialogue_acts/aggregation-results.ts
@@ -22,8 +22,8 @@ import { Ast, } from 'thingtalk';
 
 import * as C from '../../templates/ast_manip';
 
+import { ContextInfo } from '../context-info';
 import {
-    ContextInfo,
     makeAgentReply,
     makeSimpleState,
 } from '../state_manip';
