+++ lib/transaction-dialogues/dialogue_acts/coref-actions.ts
@@ -24,7 +24,7 @@ import assert from 'assert';
 import { Ast, } from 'thingtalk';
 
 import * as C from '../../templates/ast_manip';
-import { ContextInfo } from '../state_manip';
+import { ContextInfo } from '../context-info';
 
 function contextualAction(ctx : ContextInfo, action : Ast.Invocation) {
     assert(action instanceof Ast.Invocation);
