+++ lib/transaction-dialogues/dialogue_acts/coref-questions.ts
@@ -24,8 +24,8 @@ import assert from 'assert';
 import { Ast, } from 'thingtalk';
 
 import * as C from '../../templates/ast_manip';
+import { ContextInfo } from '../context-info';
 import {
-    ContextInfo,
     addQuery,
 } from '../state_manip';
 import {
