+++ lib/transaction-dialogues/dialogue_acts/refinement.ts
@@ -23,8 +23,8 @@ import { Ast, } from 'thingtalk';
 
 import * as C from '../../templates/ast_manip';
 
+import { ContextInfo } from '../context-info';
 import {
-    ContextInfo,
     makeAgentReply,
     addQuery,
 } from '../state_manip';
