+++ lib/transaction-dialogues/dialogue_acts/streams.ts
@@ -22,8 +22,8 @@ import { Ast, } from 'thingtalk';
 
 import * as C from '../../templates/ast_manip';
 
+import { ContextInfo } from '../context-info';
 import {
-    ContextInfo,
     addNewStatement,
 } from '../state_manip';
