+++ lib/transaction-dialogues/templates/recommendation.genie
@@ -32,7 +32,6 @@ import { Ast, } from 'thingtalk';
 import * as C from '../../templates/ast_manip';
 import ThingpediaLoader from '../../templates/load-thingpedia';
 import * as D from '../dialogue_acts';
-import * as S from '../state_manip';
 
 import { SlotBag } from '../../templates/slot_bag';
