+++ lib/utils/thingtalk/state-manipulation.ts
@@ -1,0 +1,41 @@
+// -*- mode: typescript; indent-tabs-mode: nil; js-basic-offset: 4 -*-
+//
+// This file is part of Genie
+//
+// Copyright 2020-2021 The Board of Trustees of the Leland Stanford Junior University
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Author: Giovanni Campagna <gcampagn@cs.stanford.edu>
+
+import { Ast } from 'thingtalk';
+
+/**
+ * Construct an agent or user target that only change the dialogue act.
+ *
+ * The resulting agent target carries over all the accepted items and not executed items
+ * from the current state.
+ */
+ export function makeSimpleTargetState(state : Ast.DialogueState|null, policyName : string, dialogueAct : string, dialogueActParam : Array<string|Ast.Value> = []) {
+    const newState = new Ast.DialogueState(null, policyName, dialogueAct, dialogueActParam, []);
+    if (state === null)
+        return newState;
+
+    for (let i = 0; i < state.history.length; i++) {
+        if (state.history[i].confirm === 'proposed' || state.history[i].results !== null)
+            continue;
+        newState.history.push(state.history[i]);
+    }
+
+    return newState;
+}
